requirements.txt
pandas
ta
binance
python-dotenv
flask
python-binance